Get follows

Get follows

get_follows(
  from_id = NULL,
  to_id = NULL,
  after = NULL,
  first = NULL,
  clean_json = TRUE
)

Arguments

from_id

A character. User ID. The request returns information about users who are being followed by the from_id user.

to_id

A character. User ID. The request returns information about users who are following the to_id user.

after

A character. Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.

first

A numeric. Maximum number of objects to return. Maximum: 100. Default: 20.

clean_json

A logical. If TRUE, clean and tidy the data. If FALSE, return the result of httr::content.

Value

A tibble data frame of follower data.

References

https://dev.twitch.tv/docs/api/reference#get-users-follows

See also

Examples

if (FALSE) {
library(twitchr)

twitch_auth()

user <- get_users(login = "KowAndToilet")

followers <- get_follows(to_id = user$id)
}