Get users

Gets information about one or more specified Twitch users. Users are identified by optional user IDs and/or login name. If neither a user ID nor a login name is specified, the user is looked up by Bearer token. Note: This package does not support the bearer token at this time.

get_users(id = NULL, login = NULL, clean_json = TRUE)

Arguments

id

A character. Optional. User ID. Multiple user IDs can be specified. Limit: 100.

login

A character. Optional. User login name. Multiple login names can be specified. Limit: 100.

clean_json

A logical. If TRUE, clean and tidy the data. If FALSE, return the result of httr::content.

Value

A tibble data frame of twitch user data.

References

https://dev.twitch.tv/docs/api/reference#get-users

See also

Examples

if (FALSE) {
library(twitchr)

twitch_auth()

users <- get_users(login = "KowAndToilet")
}