Search for Games or Categories

Returns a list of games or categories that match the query via name either entirely or partially.

search_categories(query = NULL, first = NULL, after = NULL, clean_json = TRUE)

Arguments

query

A character. URl encoded search query.

first

A numeric. Maximum number of objects to return. Maximum: 100. Default: 20.

after

Cursor for forward pagination: tells the server where to start fetching the next set of results, in a multi-page response. The cursor value specified here is from the pagination response field of a prior query.

clean_json

A logical. If TRUE, clean and tidy the data. If FALSE, return the result of httr::content.

Value

A tibble data frame of search categories data.

References

https://dev.twitch.tv/docs/api/reference#search-categories

See also

Examples

if (FALSE) {
library(twitchr)

twitch_auth()

search_results <- search_channels("fort")
}